chore(eslint-config): update dependencies and peerDependencies
Updated various dependencies and peerDependencies in the `eslint-config` package to their newer versions. This includes major version bumps for `typescript` to `>=5.0.0`, `@typescript-eslint/eslint-plugin` and `@typescript-eslint/parser` to `7.1.0`, and minor or patch updates for other plugins like `eslint-plugin-html`, `eslint-plugin-jsdoc`, `eslint-plugin-sonarjs`, `eslint-plugin-unicorn`, `eslint-plugin-vitest`, `eslint-plugin-vue`, `@types/eslint`, and `eslint` itself. This ensures that the package stays up-to-date with the latest improvements and fixes provided by these dependencies.