feat(eslint-config): enforce strict JSDoc tag names
Added a new rule to the `eslint-config` package to enforce a strict set of tags for JSDoc comments. This enhancement includes custom tags specific to our projects, improving documentation consistency and quality. The `jsdoc/check-tag-names` rule has been configured to error on undefined tags, with `category` added as an initially defined tag. This change aims to standardize our JSDoc usage across the codebase.