chore(fs): moved work in progress to new branch feat/fs
Moved several work-in-progress modules from the main branch to the `feat/fs` branch. This includes new functionality for filesystem watching, resolving aliases and imports, and handling module and type resolution. The move is part of an effort to keep the main branch clean while allowing for focused development on these features in a separate branch. The newly staged files include implementations for `FSWatcherDeep`, `resolveAlias`, `resolveImport`, `resolveModule`, `resolveTypes`, and an empty `updateObject` module.